% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3wordcloud.R
\name{d3wordcloud}
\alias{d3wordcloud}
\title{D3 JavaScript Wordcloud Graphs from R}
\usage{
d3wordcloud(words, freqs, font = "Open Sans", font.weight = 400,
  padding = 1, rotate.min = -30, rotate.max = 30, scale = "linear",
  spiral = "archimedean", width = NULL, height = NULL)
}
\description{
Create D3 JavaScript wordcloud
}
\examples{
library("tm")
library("dplyr")
data(crude)
crude <- tm_map(crude, removePunctuation)
crude <- tm_map(crude, function(x){ removeWords(x, stopwords()) })
tdm <- TermDocumentMatrix(crude)
m <- as.matrix(tdm)
v <- sort(rowSums(m), decreasing = TRUE)
d <- data.frame(word = names(v), freq = v)
d <- d \%>\% tbl_df()
d <- d \%>\% arrange(desc(freq))
d <- d \%>\% head(500)
d3wordcloud(d$word, d$freq, font = "Erica One", padding = 5)
d3wordcloud(d$word, d$freq, font = "Open Sans", padding = 5)
d3wordcloud(d$word, d$freq, font = "Open Sans", padding = 1)
d3wordcloud(d$word, d$freq, font = "Open Sans", padding = 1, font.weight = 500)
d3wordcloud(d$word, d$freq, font = "Open Sans", padding = 1, font.weight = 800)
}

