% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3wordcloud.R
\name{d3wordcloudOutput}
\alias{d3wordcloudOutput}
\title{Widget output function for use in Shiny}
\usage{
d3wordcloudOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

